Get the OS calls to be able to push the display list to the video module

+++ src/nxemu-os/core/hle/service/nvdrv/devices/nvdisp_disp0.cpp
@@ -87,7 +87,7 @@ void nvdisp_disp0::Composite(std::span<const Nvnflinger::HwcLayer> sorted_layers
         }
     }
 
-    system.GetVideo().RequestComposite(output_layers.data(), output_layers.size(), output_fences.data(), output_fences.size());
+    system.GetVideo().RequestComposite(output_layers.data(), (uint32_t)output_layers.size(), output_fences.data(), (uint32_t)output_fences.size());
     system.SpeedLimiter().DoSpeedLimiting(system.CoreTiming().GetGlobalTimeUs());
     system.GetPerfStats().EndSystemFrame();
     system.GetPerfStats().BeginSystemFrame();
